added the intersection set to test_sets
